delete confirm

+++ src/components/Inventory.js
@@ -11,6 +11,10 @@ const Inventory = () => {
     }, []);
     const keys = ['name', 'rating', 'price', 'quantity', 'supplier', 'sold'];
     const deleteItem = (id) => {
+        const ans = window.confirm(
+            'Are you sure you want to delete this item?'
+        );
+        if (!ans) return;
         fetch(`${process.env.REACT_APP_BACK_URL}/items/${id}`, {
             method: 'DELETE',
         })
